Refactor: Remove hardcoded `torch.bfloat16` cast in attention

+++ src/diffusers/models/transformers/transformer_skyreels_v2.py
@@ -105,9 +105,9 @@ class SkyReelsV2AttnProcessor2_0:
             hidden_states_img = hidden_states_img.type_as(query)
 
         hidden_states = F.scaled_dot_product_attention(
-            query.to(torch.bfloat16),
-            key.to(torch.bfloat16),
-            value.to(torch.bfloat16),
+            query,
+            key,
+            value,
             attn_mask=attention_mask if self._flag_ar_attention else None,
             dropout_p=0.0,
             is_causal=False,
